Change Health
No longer extended from Bounded, but includes bounded as a value. I didn't want things like health:setValue, and I didn't want to have to get rid of them because then you end up with that weird issue that java tries to fix with interfaces where there like "what methods does this actually include". Health is health, and is not the same as a boundedstat.
also added a fullheal method and some comments that explain the return values